+++ app/models/isonline/his_jhcis.model.js
@@ -13,10 +13,20 @@ class HisJhcisModel {
             .where('table_schema', '=', dbName);
     }
     async testConnect(db) {
-        const result = await global.dbHIS('j2_hospital').first();
+        let result;
+        result = await global.dbHIS('j2_hospital').first();
         const hospname = result?.HNAME || null;
-        const patient = await db('person').select('pid as hn').limit(1);
-        return { hospname, patient };
+        result = await db('person').select('pid as hn').limit(1);
+        const connection = result && (result.patient || result.length > 0) ? true : false;
+        let charset = '';
+        if (process.env.HIS_DB_CLIENT.includes('mysql')) {
+            result = await db('information_schema.SCHEMATA')
+                .select('DEFAULT_CHARACTER_SET_NAME')
+                .where('SCHEMA_NAME', process.env.HIS_DB_NAME)
+                .first();
+            charset = result?.DEFAULT_CHARACTER_SET_NAME || '';
+        }
+        return { hospname, connection, charset };
     }
     getPerson(knex, columnName, searchText) {
         columnName = columnName === 'cid' ? 'idcard' : columnName;