+++ src/models/isonline/his_hosxppcu.model.ts
@@ -10,10 +10,22 @@ export class HisHosxppcuModel {
     }
 
     async testConnect(db: Knex) {
-        const result = await global.dbHIS('opdconfig').first();
+        let result: any;
+        result = await global.dbHIS('opdconfig').first();
         const hospname = result?.hospitalcode || null;
-        const patient = await db('patient').select('hn').limit(1);
-        return { hospname, patient }
+
+        result = await db('patient').select('hn').limit(1);
+        const connection = result && (result.patient || result.length > 0) ? true : false;
+
+        let charset: any = '';
+        if (process.env.HIS_DB_CLIENT.includes('mysql')) {
+            result = await db('information_schema.SCHEMATA')
+                .select('DEFAULT_CHARACTER_SET_NAME')
+                .where('SCHEMA_NAME', process.env.HIS_DB_NAME)
+                .first();
+            charset = result?.DEFAULT_CHARACTER_SET_NAME || '';
+        }
+        return { hospname, connection, charset };
     }
 
     getPerson(db: Knex, columnName: string, searchText: any) {