+++ src/middleware/moph-refer.ts
@@ -134,7 +134,7 @@ export const checkAdminRequest = async () => {
     return { status: 500, message: 'No nRefer token' };
   }
 
-  const url = referAPIUrl + '/his-connect/checkRequest/'+hcode;
+  const url = referAPIUrl + '/moph-erp/check-request/'+hcode;
   const headers = {
     'Content-Type': 'application/json',
     'client-ip': apiIp.ip,